Move OpenSSL helper objects to own files
