+++ sdk/main/src/impl/OpenSSLUtils.cc
@@ -17,9 +17,10 @@
  * limitations under the License.
  *
  */
-#include "impl/OpenSSLUtils.h"
+#include "impl/openssl_utils/OpenSSLUtils.h"
 #include "exceptions/OpenSSLException.h"
-#include "impl/HexConverter.h"
+#include "impl/openssl_utils/EVP_MD.h"
+#include "impl/openssl_utils/EVP_MD_CTX.h"
 
 #include <openssl/err.h>
 #include <openssl/hmac.h>
@@ -42,124 +43,6 @@ constexpr const size_t ERROR_MSG_SIZE = 256ULL;
 }
 
 //-----
-BIGNUM::BIGNUM(::BIGNUM* bignum)
-  : OpenSSLObjectWrapper(bignum, &BN_clear_free, &BN_dup)
-{
-}
-
-//-----
-BIGNUM BIGNUM::fromHex(std::string_view hexString)
-{
-  ::BIGNUM* bigNum = nullptr;
-  if (BN_hex2bn(&bigNum, hexString.data()) <= 0)
-  {
-    throw OpenSSLException(getErrorMessage("BN_hex2bn"));
-  }
-
-  return BIGNUM(bigNum);
-}
-
-//-----
-BIGNUM BIGNUM::fromBytes(const std::vector<unsigned char>& bytes)
-{
-  // go through hex rather than using big-endian openssl functions
-  return fromHex(HexConverter::bytesToHex(bytes));
-}
-
-//-----
-BIGNUM BIGNUM::modularAdd(const BIGNUM& other, const BIGNUM& modulo) const
-{
-  const BN_CTX context(BN_CTX_secure_new());
-  if (!context)
-  {
-    throw OpenSSLException(getErrorMessage("BN_CTX_secure_new"));
-  }
-
-  BIGNUM result(BN_secure_new());
-  if (!result)
-  {
-    throw OpenSSLException(getErrorMessage("BN_secure_new"));
-  }
-
-  if (BN_mod_add(result.get(), get(), other.get(), modulo.get(), context.get()) <= 0)
-  {
-    throw OpenSSLException(getErrorMessage("BN_mod_add"));
-  }
-
-  return result;
-}
-
-//-----
-std::vector<unsigned char> BIGNUM::toBytes() const
-{
-  char* hex = BN_bn2hex(get());
-  const std::string hexString(hex);
-  OPENSSL_free(hex);
-
-  return HexConverter::hexToBytes(hexString);
-}
-
-//-----
-BN_CTX::BN_CTX(::BN_CTX* bnCtx)
-  : OpenSSLObjectWrapper(bnCtx, &BN_CTX_free)
-{
-}
-
-//-----
-EC_GROUP::EC_GROUP(::EC_GROUP* ecGroup)
-  : OpenSSLObjectWrapper(ecGroup, &EC_GROUP_free, &EC_GROUP_dup)
-{
-}
-
-//-----
-EC_POINT::EC_POINT(::EC_POINT* ecPoint)
-  : OpenSSLObjectWrapper(ecPoint, &EC_POINT_free)
-{
-}
-
-//-----
-ECDSA_SIG::ECDSA_SIG(::ECDSA_SIG* ecdsaSig)
-  : OpenSSLObjectWrapper(ecdsaSig, &ECDSA_SIG_free)
-{
-}
-
-//-----
-EVP_MD::EVP_MD(::EVP_MD* evpMd)
-  : OpenSSLObjectWrapper(evpMd, &EVP_MD_free)
-{
-}
-
-//-----
-EVP_MD_CTX::EVP_MD_CTX(::EVP_MD_CTX* evpMdCtx)
-  : OpenSSLObjectWrapper(evpMdCtx, &EVP_MD_CTX_free)
-{
-}
-
-//-----
-EVP_PKEY::EVP_PKEY(::EVP_PKEY* evpPkey)
-  : OpenSSLObjectWrapper(evpPkey, &EVP_PKEY_free, &EVP_PKEY_dup)
-{
-}
-
-//-----
-EVP_PKEY_CTX::EVP_PKEY_CTX(::EVP_PKEY_CTX* evpPkeyCtx)
-  : OpenSSLObjectWrapper(evpPkeyCtx, &EVP_PKEY_CTX_free, &EVP_PKEY_CTX_dup)
-{
-}
-
-//-----
-OSSL_LIB_CTX::OSSL_LIB_CTX(::OSSL_LIB_CTX* osslLibCtx)
-  : OpenSSLObjectWrapper(osslLibCtx, &OSSL_LIB_CTX_free)
-{
-}
-
-//-----
-OSSL_DECODER_CTX::OSSL_DECODER_CTX(::OSSL_DECODER_CTX* osslDecoderCtx)
-  : OpenSSLObjectWrapper(osslDecoderCtx, &OSSL_DECODER_CTX_free)
-{
-}
-
-//-----
 std::vector<unsigned char> computeSHA256(const std::vector<unsigned char>& data)
 {
   auto outputBytes = std::vector<unsigned char>(SHA256_HASH_SIZE);
@@ -179,7 +62,7 @@ std::vector<unsigned char> computeSHA384(const std::vector<unsigned char>& data)
 std::vector<unsigned char> computeSHA512HMAC(const std::vector<unsigned char>& key,
                                              const std::vector<unsigned char>& data)
 {
-  const EVP_MD_CTX messageDigestContext(EVP_MD_CTX_new());
+  EVP_MD_CTX messageDigestContext(EVP_MD_CTX_new());
   if (!messageDigestContext)
   {
     throw OpenSSLException(getErrorMessage("EVP_MD_CTX_new"));
